requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
msal==1.31.0
httpx==0.28.1
google-api-python-client==2.111.0
google-auth==2.27.0
